moved __version__ variable to starling_sim/__init__.py
