--- starling_sim/version.py
@@ -1,3 +1,0 @@
-"""Starling version"""
-
-__version__ = "0.10.0"
